`ArticleService`と`UserService`をドメイン層へ移動
