+++ src/domain/models/mod.rs
@@ -2,3 +2,5 @@ pub mod article;
 pub mod article_query;
 pub mod user;
 pub mod user_name;
+pub mod article_service;
+pub mod user_service;
