+++ src/usecase/article_usecase.rs
@@ -1,7 +1,7 @@
 use async_trait::async_trait;
 use mongodb::bson::oid::ObjectId;
 
-use crate::domain::{models::{article::Article, user_name::UserName}, repositorys::article_repository::ArticleRepository};
+use crate::domain::{models::{article::Article, user_name::UserName, article_service::ArticleService}, repositorys::article_repository::ArticleRepository};
 use crate::domain::models::article_query::ArticleQuery;
 
 #[derive(Clone)]
@@ -13,40 +13,6 @@ impl<T: ArticleRepository + Clone> ArticleUsecase<T> {
     pub fn new(repository: T) -> Self {
         ArticleUsecase { repository }
     }
-}
-
-
-
-#[async_trait]
-pub trait ArticleService {
-    async fn get_articles(
-        &self,
-        skip: usize,
-        limit: usize,
-    ) -> Result<Vec<Article>, mongodb::error::Error>;
-    async fn get_article_by_id(
-        &self,
-        id: ObjectId,
-    ) -> Result<Option<Article>, mongodb::error::Error>;
-    async fn create_article(
-        &self,
-        title: String,
-        author: UserName,
-        content: String,
-    ) -> Result<Article, mongodb::error::Error>;
-    async fn update_article(
-        &self,
-        id: ObjectId,
-        title: Option<String>,
-        content: Option<String>,
-    ) -> Result<Article, mongodb::error::Error>;
-    async fn delete_article(&self, id: ObjectId) -> Result<(), mongodb::error::Error>;
-    async fn search_articles(
-        &self,
-        skip: usize,
-        limit: usize,
-        query: ArticleQuery,
-    ) -> Result<Vec<Article>, mongodb::error::Error>;
 }
 
 #[async_trait]
