+++ src/usecase/user_usecase.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 use sha2::{Digest, Sha256};
 
 use crate::domain::{
-    models::{user::User, user_name::UserName},
+    models::{user::User, user_name::UserName, user_service::UserService},
     repositorys::user_repository::UserRepository,
 };
 
@@ -15,36 +15,6 @@ impl<T: UserRepository + Clone> UserUsecase<T> {
     pub fn new(repository: T) -> Self {
         UserUsecase { repository }
     }
-}
-
-#[async_trait]
-pub trait UserService {
-    async fn get_users(
-        &self,
-        skip: usize,
-        limit: usize,
-    ) -> Result<Vec<User>, mongodb::error::Error>;
-    async fn get_user_by_name(&self, name: &str) -> Result<Option<User>, mongodb::error::Error>;
-    async fn create_user(
-        &self,
-        name: String,
-        display_name: String,
-        intro: String,
-        email: String,
-        show_email: bool,
-        password: String,
-    ) -> Result<User, mongodb::error::Error>;
-    async fn update_user(
-        &self,
-        name: String,
-        display_name: Option<String>,
-        intro: Option<String>,
-        email: Option<String>,
-        show_email: Option<bool>,
-        password: Option<String>,
-    ) -> Result<User, mongodb::error::Error>;
-    async fn delete_user(&self, name: &str) -> Result<(), mongodb::error::Error>;
-    async fn validate_user_name(&self, name: &str) -> Result<UserName, mongodb::error::Error>;
 }
 
 #[async_trait]
